# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User has been successfully created"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "summary": "Create a new user",
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns a list of users"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all users (Admin only)",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me": {
      "get": {
        "operationId": "UsersController_getCurrentUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns the current user profile"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the currently authenticated user",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "example": "60d21b4667d0d8992e610c85",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the user information",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "60d21b4667d0d8992e610c85",
                    "email": "user@example.com",
                    "name": "John Doe",
                    "role": "streamer",
                    "image": "https://example.com/avatar.jpg"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user by ID",
        "tags": [
          "users"
        ]
      },
      "put": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "example": "60d21b4667d0d8992e610c85",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User has been successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "60d21b4667d0d8992e610c85",
                    "email": "updated@example.com",
                    "name": "Updated Name",
                    "role": "streamer"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Cannot update other users unless admin"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user information",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/profile/streamer/{id}": {
      "get": {
        "operationId": "UsersController_getStreamerProfile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Streamer ID",
            "schema": {
              "example": "60d21b4667d0d8992e610c85",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the streamer profile",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "60d21b4667d0d8992e610c85",
                    "name": "Popular Streamer",
                    "image": "https://example.com/avatar.jpg",
                    "channelUrl": "https://twitch.tv/popularstreamer",
                    "category": [
                      "Gaming"
                    ],
                    "language": [
                      "English"
                    ],
                    "statistics": {
                      "followers": 10000,
                      "averageViewers": 500,
                      "impressions": 25000
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Streamer not found"
          }
        },
        "summary": "Get streamer profile information",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me/overlay": {
      "get": {
        "operationId": "UsersController_getOverlaySettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      },
      "put": {
        "operationId": "UsersController_updateOverlaySettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OverlaySettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Overlay settings successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverlaySettingsResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update overlay settings for the current user",
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/v1/users/me/overlay/status": {
      "get": {
        "operationId": "UsersController_getOverlayStatus",
        "parameters": [
          {
            "name": "overlayToken",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/v1/users/me/campaign-selection": {
      "get": {
        "operationId": "UsersController_getCampaignSelectionSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Campaign selection settings retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignSelectionSettingsResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get campaign selection settings for the current user",
        "tags": [
          "users"
        ]
      },
      "put": {
        "operationId": "UsersController_updateCampaignSelectionSettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignSelectionSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Campaign selection settings successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignSelectionSettingsResponseDto"
                }
              }
            }
          }
        },
        "summary": "Update campaign selection settings for the current user",
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/v1/users/me/streak": {
      "get": {
        "operationId": "UsersController_getMyStreak",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user streak summary",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me/energy-packs": {
      "get": {
        "operationId": "UsersController_getEnergyPacks",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Energy packs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnergyPacksResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user energy packs",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me/energy-packs/consume": {
      "post": {
        "operationId": "UsersController_consumeEnergyPack",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumeEnergyPackDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Energy pack consumed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "remaining": 9
                  }
                }
              }
            }
          },
          "400": {
            "description": "No energy packs available"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Consume one energy pack",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me/xp": {
      "get": {
        "operationId": "UsersController_getXP",
        "parameters": [],
        "responses": {
          "200": {
            "description": "XP data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/XPResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user XP data",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_addXP",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddXPDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "XP added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/XPResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid activity type or amount"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add XP for an activity",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me/rp": {
      "get": {
        "operationId": "UsersController_getMyRP",
        "parameters": [],
        "responses": {
          "200": {
            "description": "RP data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user RP data",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_addRP",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddRPDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "RP added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid activity type or amount"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add RP for an activity",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me/level": {
      "get": {
        "operationId": "UsersController_getMyLevel",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Level data retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user level data",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me/level/check": {
      "post": {
        "operationId": "UsersController_checkMyLevelUp",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Level check completed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Check for level up and update if necessary",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "GameTriggers API",
    "description": "The Gametriggers marketplace API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UserRole": {
        "type": "string",
        "enum": [
          "streamer",
          "brand",
          "admin"
        ],
        "description": "User role in the system"
      },
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "user@example.com"
          },
          "name": {
            "type": "string",
            "description": "User full name",
            "example": "John Doe"
          },
          "image": {
            "type": "string",
            "description": "User profile image URL",
            "example": "https://example.com/avatar.jpg"
          },
          "role": {
            "description": "User role in the system",
            "example": "streamer",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "authProvider": {
            "description": "Authentication provider",
            "example": "twitch",
            "allOf": [
              {
                "$ref": "#/components/schemas/AuthProvider"
              }
            ]
          },
          "authProviderId": {
            "type": "string",
            "description": "ID from the authentication provider",
            "example": "12345678"
          },
          "channelUrl": {
            "type": "string",
            "description": "Channel URL for streamers",
            "example": "https://twitch.tv/username"
          },
          "category": {
            "description": "Stream categories",
            "example": [
              "Gaming",
              "Just Chatting"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "language": {
            "description": "Stream languages",
            "example": [
              "English",
              "Spanish"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "email",
          "name",
          "role",
          "authProvider"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {}
      },
      "OverlaySettingsDto": {
        "type": "object",
        "properties": {
          "position": {
            "type": "string",
            "description": "Position of the overlay on screen",
            "enum": [
              "top-left",
              "top-right",
              "bottom-left",
              "bottom-right"
            ],
            "example": "top-right"
          },
          "size": {
            "type": "string",
            "description": "Size of the overlay",
            "enum": [
              "small",
              "medium",
              "large"
            ],
            "example": "medium"
          },
          "opacity": {
            "type": "object",
            "description": "Opacity percentage of the overlay",
            "minimum": 20,
            "maximum": 100,
            "example": 75
          },
          "backgroundColor": {
            "type": "object",
            "description": "Background color in hex or rgba format",
            "example": "#ffffff"
          }
        }
      },
...
      "ConsumeEnergyPackDto": {
        "type": "object",
        "properties": {
          "campaignId": {
            "type": "string",
            "description": "Campaign ID that energy pack is being used for",
            "example": "60d21b4667d0d8992e610c85"
          }
        },
        "required": [
          "campaignId"
        ]
      },
...
    }
  }
}